requirements.txt
altgraph==0.17.4
bidict==0.23.1
blinker==1.9.0
cffi==1.17.1
click==8.2.1
dnspython==2.7.0
eventlet==0.40.2
Flask==3.1.1
Flask-SocketIO==5.5.1
greenlet==3.2.3
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
macholib==1.16.3
MarkupSafe==3.0.2
numpy==2.3.2
packaging==25.0
pycparser==2.22
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
setuptools==80.9.0
simple-websocket==1.1.0
sounddevice==0.5.2
Werkzeug==3.1.3
wsproto==1.2.0
